Data.Geometry.Ipe.Attributes

Description

Possible Attributes we can assign to items in an Ipe file

Documentation

data AttributeUniverse

Constructors

Layer

Matrix

Pin

Transformations

Stroke

Fill

Pen

Size

Dash

LineCap

LineJoin

FillRule

Arrow

RArrow

Opacity

Tiling

Gradient

Clip

type LayerSym0 = 'Layer

type MatrixSym0 = 'Matrix

type PinSym0 = 'Pin

type TransformationsSym0 = 'Transformations

type StrokeSym0 = 'Stroke

type FillSym0 = 'Fill

type PenSym0 = 'Pen

type SizeSym0 = 'Size

type DashSym0 = 'Dash

type LineCapSym0 = 'LineCap

type LineJoinSym0 = 'LineJoin

type FillRuleSym0 = 'FillRule

type ArrowSym0 = 'Arrow

type RArrowSym0 = 'RArrow

type OpacitySym0 = 'Opacity

type TilingSym0 = 'Tiling

type GradientSym0 = 'Gradient

type ClipSym0 = 'Clip

data SAttributeUniverse :: AttributeUniverse -> Type where

Constructors

SLayer :: SAttributeUniverse 'Layer

SMatrix :: SAttributeUniverse 'Matrix

SPin :: SAttributeUniverse 'Pin

STransformations :: SAttributeUniverse 'Transformations

SStroke :: SAttributeUniverse 'Stroke

SFill :: SAttributeUniverse 'Fill

SPen :: SAttributeUniverse 'Pen

SSize :: SAttributeUniverse 'Size

SDash :: SAttributeUniverse 'Dash

SLineCap :: SAttributeUniverse 'LineCap

SLineJoin :: SAttributeUniverse 'LineJoin

SFillRule :: SAttributeUniverse 'FillRule

SArrow :: SAttributeUniverse 'Arrow

SRArrow :: SAttributeUniverse 'RArrow

SOpacity :: SAttributeUniverse 'Opacity

STiling :: SAttributeUniverse 'Tiling

SGradient :: SAttributeUniverse 'Gradient

SClip :: SAttributeUniverse 'Clip

type CommonAttributes = [Layer, Matrix, Pin, Transformations]

type TextLabelAttributes = CommonAttributes

type MiniPageAttributes = CommonAttributes

type ImageAttributes = CommonAttributes

type SymbolAttributes = CommonAttributes ++ [Stroke, Fill, Pen, Size]

type PathAttributes = CommonAttributes ++ [Stroke, Fill, Dash, Pen, LineCap, LineJoin, FillRule, Arrow, RArrow, Opacity, Tiling, Gradient]

type GroupAttributes = CommonAttributes ++ '['Clip]

Attr

newtype Attr (f :: TyFun u * -> *) (label :: u)

Attr implements the mapping from labels to types as specified by the (symbol representing) the type family f

Constructors

GAttr

Fields

_getAttr :: Maybe (Apply f label)

getAttr :: forall f label f label. Iso (Attr f label) (Attr f label) (Maybe (Apply f label)) (Maybe (Apply f label))

pattern Attr :: Apply f label -> Attr f label

Constructor for constructing an Attr given an actual value.

pattern NoAttr :: Attr f label

An Attribute that is not set

traverseAttr :: Applicative h => (Apply f label -> h (Apply g label)) -> Attr f label -> h (Attr g label)

pureAttr :: (Applicative h, Apply f a ~ Apply g a) => Attr f a -> h (Attr g a)

Traverse for the situation where the type is not actually parameterized.

Attributes

newtype Attributes (f :: TyFun u * -> *) (ats :: [u])

A collection of Attributes.

Constructors

Attrs (Rec (Attr f) ats)

unAttrs :: Lens (Attributes f ats) (Attributes f' ats') (Rec (Attr f) ats) (Rec (Attr f') ats')

traverseAttrs :: Applicative h => (forall label. Attr f label -> h (Attr g label)) -> Attributes f ats -> h (Attributes g ats)

zipRecsWith :: (forall a. f a -> g a -> h a) -> Rec f as -> Rec g as -> Rec h as

ixAttr :: forall at ats proxy f. at ∈ ats => proxy at -> Lens' (Attributes f ats) (Maybe (Apply f at))

Lens into a specific attribute, if it is set.

_Attr :: forall at ats proxy f. (at ∈ ats, RecApplicative ats) => proxy at -> Prism' (Attributes f ats) (Apply f at)

Prism into a particular attribute.

lookupAttr :: at ∈ ats => proxy at -> Attributes f ats -> Maybe (Apply f at)

Looks up a particular attribute.

setAttr :: forall proxy at ats f. at ∈ ats => proxy at -> Apply f at -> Attributes f ats -> Attributes f ats

Sets a particular attribute

takeAttr :: forall proxy at ats f. at ∈ ats => proxy at -> Attributes f ats -> (Maybe (Apply f at), Attributes f ats)

gets and removes the attribute from Attributes

unSetAttr :: forall proxy at ats f. at ∈ ats => proxy at -> Attributes f ats -> Attributes f ats

unsets/Removes an attribute

attr :: (at ∈ ats, RecApplicative ats) => proxy at -> Apply f at -> Attributes f ats

Creates a singleton attribute

data PinType

Common Attributes

Possible values for Pin

Constructors

No

Yes

Horizontal

Vertical

data TransformationTypes

Possible values for Transformation

Constructors

Affine

Rigid

Translations

newtype IpeSize r

TODO

Symbol Attributes

The optional Attributes for a symbol data SymbolAttributeUniverse = SymbolStroke | SymbolFill | SymbolPen | Size deriving (Show,Eq)

Constructors

IpeSize (IpeValue r)

newtype IpePen r

Constructors

IpePen (IpeValue r)

data IpeDash r

Path Attributes

Possible attributes for a path data PathAttributeUniverse = Stroke | Fill | Dash | Pen | LineCap | LineJoin | FillRule | Arrow | RArrow | Opacity | Tiling | Gradient deriving (Show,Eq)

Possible values for Dash

Constructors

DashNamed Text

DashPattern [r] r

data FillType

Allowed Fill types

Constructors

Wind

EOFill

type IpeOpacity = Text

IpeOpacity, IpeTyling, and IpeGradient are all symbolic values

type IpeTiling = Text

type IpeGradient = Text

data IpeArrow r

Possible values for an ipe arrow

Constructors

IpeArrow

Fields

_arrowName :: Text

_arrowSize :: IpeSize r

arrowSize :: forall r r. Lens (IpeArrow r) (IpeArrow r) (IpeSize r) (IpeSize r)

arrowName :: forall r. Lens' (IpeArrow r) Text

normalArrow :: IpeArrow r

Attribute names in Ipe

class IpeAttrName (a :: AttributeUniverse) where

For the types representing attribute values we can get the name/key to use when serializing to ipe.

Methods

attrName :: proxy a -> Text

writeAttrNames :: AllConstrained IpeAttrName rs => Rec f rs -> Rec (Const Text) rs

Writing Attribute names