Data.Geometry.Ipe.Color

Description

Data type for representing colors in ipe as well as the colors available in the standard ipe stylesheet.

Documentation

newtype IpeColor r

Constructors

IpeColor (IpeValue (RGB r))

named :: Text -> IpeColor r

Creates a named color

Basic Named colors

red :: IpeColor r

green :: IpeColor r

blue :: IpeColor r

yellow :: IpeColor r

orange :: IpeColor r

gold :: IpeColor r

purple :: IpeColor r

gray :: IpeColor r

brown :: IpeColor r

navy :: IpeColor r

pink :: IpeColor r

seagreen :: IpeColor r

turquoise :: IpeColor r

violet :: IpeColor r

darkblue :: IpeColor r

darkcyan :: IpeColor r

darkgray :: IpeColor r

darkgreen :: IpeColor r

darkmagenta :: IpeColor r

darkorange :: IpeColor r

darkred :: IpeColor r

lightblue :: IpeColor r

lightcyan :: IpeColor r

lightgray :: IpeColor r

lightgreen :: IpeColor r

lightyellow :: IpeColor r