Data.Geometry.Ipe.FromIpe

Description

Functions that help reading geometric values from ipe images.

Individual readers

_asPoint :: Prism' (IpeSymbol r) (Point 2 r)

Extracts the point from a Symbol. When creating a symbol this creates a disk that supports a stroke color.

_asLineSegment :: Prism' (Path r) (LineSegment 2 () r)

Try to convert a path into a line segment, fails if the path is not a line segment or a polyline with more than two points.

_asRectangle :: forall r. (Num r, Ord r) => Prism' (Path r) (Rectangle () r)

Tries to convert a path into a rectangle.

_asTriangle :: Prism' (Path r) (Triangle 2 () r)

Convert to a triangle

_asPolyLine :: Prism' (Path r) (PolyLine 2 () r)

Convert to a polyline. Ignores all non-polyline parts

>>> testPath ^? _asPolyLine
Just (PolyLine {_points = LSeq (fromList [Point2 [0,0] :+ (),Point2 [10,10] :+ (),Point2 [200,100] :+ ()])})

_asSomePolygon :: Prism' (Path r) (SomePolygon () r)

_asSimplePolygon :: Prism' (Path r) (Polygon Simple () r)

Convert to a simple polygon

_asMultiPolygon :: Prism' (Path r) (MultiPolygon () r)

Convert to a multipolygon

Dealing with Attributes

_withAttrs :: Prism' (IpeObject r) (i r :+ IpeAttributes i r) -> Prism' (i r) g -> Prism' (IpeObject r) (g :+ IpeAttributes i r)

Use the first prism to select the ipe object to depicle with, and the second how to select the geometry object from there on. Then we can select the geometry object, directly with its attributes here.

>>> testObject ^? _withAttrs _IpePath _asPolyLine
Just (PolyLine {_points = LSeq (fromList [Point2 [0,0] :+ (),Point2 [10,10] :+ (),Point2 [200,100] :+ ()])} :+ Attrs {NoAttr, NoAttr, NoAttr, NoAttr, Attr IpeColor (Named "red"), NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr})

Default readers

class HasDefaultFromIpe g where

Associated Types

type DefaultFromIpe g :: * -> *

Methods

defaultFromIpe :: r ~ NumType g => Prism' (IpeObject r) (g :+ IpeAttributes (DefaultFromIpe g) r)

Reading all elements of a particular type

readAll :: forall g r. (HasDefaultFromIpe g, r ~ NumType g) => IpePage r -> [g :+ IpeAttributes (DefaultFromIpe g) r]

Read all g's from some ipe page(s).

readAllFrom :: (HasDefaultFromIpe g, r ~ NumType g, Coordinate r, Eq r) => FilePath -> IO [g :+ IpeAttributes (DefaultFromIpe g) r]

Convenience function from reading all g's from an ipe file. If there is an error reading or parsing the file the error is "thrown away".