Data.Geometry.Ipe.IpeOut

Description

Functions that help drawing geometric values in ipe. An IpeOut is essenitally a function that converts a geometric type g into an IpeObject.

We also proivde a HasDefaultIpeOut typeclass that defines a default conversion function from a geometry type g to an ipe type.

Documentation

>>> :set -XOverloadedStrings
>>> :{
let myPolygon = fromPoints . map ext $ [origin, Point2 10 10, Point2 100 200]
:}

The IpeOut type and the default combinator to use it

type IpeOut g i r = g -> IpeObject' i r

(!) :: IpeObject' i r -> IpeAttributes i r -> IpeObject' i r

Add attributes to an IpeObject'

iO :: ToObject i => IpeObject' i r -> IpeObject r

Render an ipe object

>>> :{
  iO $ defIO myPolygon ! attr SFill (IpeColor "blue")
                       ! attr SLayer "alpha"
                       ! attr SLayer "beta"
:}
IpePath (Path {_pathSegments = LSeq (fromList [PolygonPath SimplePolygon CSeq [Point2 [0,0] :+ (),Point2 [10,10] :+ (),Point2 [100,200] :+ ()]])} :+ Attrs {Attr LayerName {_layerName = "beta"}, NoAttr, NoAttr, NoAttr, NoAttr, Attr IpeColor (Named "blue"), NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr})

>>> :{
  iO $ ipeGroup [ iO $ ipePolygon myPolygon ! attr SFill (IpeColor "red")
                ] ! attr SLayer "alpha"
:}
IpeGroup (Group [IpePath (Path {_pathSegments = LSeq (fromList [PolygonPath SimplePolygon CSeq [Point2 [0,0] :+ (),Point2 [10,10] :+ (),Point2 [100,200] :+ ()]])} :+ Attrs {NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, Attr IpeColor (Named "red"), NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr})] :+ Attrs {Attr LayerName {_layerName = "alpha"}, NoAttr, NoAttr, NoAttr, NoAttr})

iO'' :: (HasDefaultIpeOut g, NumType g ~ r, DefaultIpeOut g ~ i, ToObject i) => g -> IpeAttributes i r -> IpeObject r

Render to an ipe object using the defIO IpeOut

>>> :{
  iO'' myPolygon $  attr SFill (IpeColor "red")
                 <> attr SLayer "alpha"
                 <> attr SLayer "beta"
:}
IpePath (Path {_pathSegments = LSeq (fromList [PolygonPath SimplePolygon CSeq [Point2 [0,0] :+ (),Point2 [10,10] :+ (),Point2 [100,200] :+ ()]])} :+ Attrs {Attr LayerName {_layerName = "beta"}, NoAttr, NoAttr, NoAttr, NoAttr, Attr IpeColor (Named "red"), NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr})

>>> iO'' [ myPolygon , myPolygon ] $ attr SLayer "alpha"
IpeGroup (Group [IpePath (Path {_pathSegments = LSeq (fromList [PolygonPath SimplePolygon CSeq [Point2 [0,0] :+ (),Point2 [10,10] :+ (),Point2 [100,200] :+ ()]])} :+ Attrs {NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr}),IpePath (Path {_pathSegments = LSeq (fromList [PolygonPath SimplePolygon CSeq [Point2 [0,0] :+ (),Point2 [10,10] :+ (),Point2 [100,200] :+ ()]])} :+ Attrs {NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr, NoAttr})] :+ Attrs {Attr LayerName {_layerName = "alpha"}, NoAttr, NoAttr, NoAttr, NoAttr})

iO' :: HasDefaultIpeOut g => g -> IpeObject (NumType g)

generate an ipe object without any specific attributes

Default Conversions

class ToObject (DefaultIpeOut g) => HasDefaultIpeOut g where

Class that specifies a default conversion from a geometry type g into an ipe object.

Associated Types

type DefaultIpeOut g :: * -> *

Methods

defIO :: IpeOut g (DefaultIpeOut g) (NumType g)

Point Converters

ipeMark :: Text -> IpeOut (Point 2 r) IpeSymbol r

ipeDiskMark :: IpeOut (Point 2 r) IpeSymbol r

Path Converters

defaultBox :: Num r => Rectangle () r

Size of the default bounding box used to clip lines and half-lines in the default IpeOuts.

ipeLine :: (Ord r, Fractional r) => IpeOut (Line 2 r) Path r

Renders a line as a Path. The line is clipped to the defaultBox

ipeLineIn :: forall p r. (Ord r, Fractional r) => Rectangle p r -> IpeOut (Line 2 r) Path r

Renders the line in the given box.

pre: the intersection of the box with the line is non-empty

ipeHalfLine :: (Ord r, Fractional r) => IpeOut (HalfLine 2 r) Path r

Renders an Halfine.

pre: the intersection of the box with the line is non-empty

ipeHalfLineIn :: forall p r. (Ord r, Fractional r) => Rectangle p r -> IpeOut (HalfLine 2 r) Path r

Renders the HalfLine in the given box.

pre: the intersection of the box with the line is non-empty

ipeLineSegment :: IpeOut (LineSegment 2 p r) Path r

ipePolyLine :: IpeOut (PolyLine 2 p r) Path r

ipeEllipse :: IpeOut (Ellipse r) Path r

ipeCircle :: Floating r => IpeOut (Circle p r) Path r

ipeDisk :: Floating r => IpeOut (Disk p r) Path r

path :: PathSegment r -> Path r

Helper to construct a path from a singleton item

pathSegment :: LineSegment 2 p r -> PathSegment r

ipePolygon :: IpeOut (Polygon t p r) Path r

Draw a polygon

ipeRectangle :: Num r => IpeOut (Rectangle p r) Path r

Draw a Rectangle

Group Converters

ipeGroup :: Foldable f => IpeOut (f (IpeObject r)) Group r

Text Converters

ipeLabel :: IpeOut (Text :+ Point 2 r) TextLabel r

Creates an text label

labelled

Arguments

:: (Show lbl, NumType g ~ r, ToObject i)

=> (g -> Point 2 r)
where to place the label

-> IpeOut g i r
how to draw the geometric object

-> IpeOut (g :+ lbl) Group r

Annotate an IpeOut with a label