Data.Geometry.Ipe.Matrix

Documentation

applyMatrix' :: (IsTransformable (i r), Matrix ∈ AttributesOf i, Dimension (i r) ~ 2, r ~ NumType (i r)) => IpeObject' i r -> IpeObject' i r

Takes and applies the ipe Matrix attribute of this item.

applyMatrix :: Fractional r => IpeObject r -> IpeObject r

Applies the matrix to an ipe object if it has one.

applyMatrices :: Fractional r => IpeFile r -> IpeFile r

applyMatricesPage :: Fractional r => IpePage r -> IpePage r