Data.Geometry.Ipe.Types

Description

Data type modeling the various elements in Ipe files.

Documentation

newtype LayerName

Constructors

LayerName Text

layerName :: Iso' LayerName Text

data Image r

Image Objects

Constructors

Image () (Rectangle () r)

imageData :: forall r. Lens' (Image r) ()

rect :: forall r r. Lens (Image r) (Image r) (Rectangle () r) (Rectangle () r)

data TextLabel r

Text Objects

Constructors

Label Text (Point 2 r)

data MiniPage r

Constructors

MiniPage Text (Point 2 r) r

width :: MiniPage t -> t

data IpeSymbol r

Ipe Symbols, i.e. Points

A symbol (point) in ipe

Constructors

Symbol (Point 2 r) Text

symbolPoint :: forall r r. Lens (IpeSymbol r) (IpeSymbol r) (Point 2 r) (Point 2 r)

symbolName :: forall r. Lens' (IpeSymbol r) Text

newtype Path r

A path is a non-empty sequence of PathSegments.

Constructors

Path (LSeq 1 (PathSegment r))

pathSegments :: forall r r. Iso (Path r) (Path r) (LSeq 1 (PathSegment r)) (LSeq 1 (PathSegment r))

data PathSegment r

Paths

Paths consist of Path Segments. PathSegments come in the following forms:

Constructors

PolyLineSegment (PolyLine 2 () r)

PolygonPath (SimplePolygon () r)

CubicBezierSegment (BezierSpline 3 2 r)

QuadraticBezierSegment (BezierSpline 2 2 r)

EllipseSegment (Ellipse r)

ArcSegment

SplineSegment

ClosedSplineSegment

newtype Group r

Groups and Objects

Group Attributes

A group is essentially a list of IpeObjects.

Constructors

Group [IpeObject r]

groupItems :: Lens (Group r) (Group s) [IpeObject r] [IpeObject s]

data IpeObject r

Constructors

IpeGroup (IpeObject' Group r)

IpeImage (IpeObject' Image r)

IpeTextLabel (IpeObject' TextLabel r)

IpeMiniPage (IpeObject' MiniPage r)

IpeUse (IpeObject' IpeSymbol r)

IpePath (IpeObject' Path r)

_IpeGroup :: forall r. Prism' (IpeObject r) (IpeObject' Group r)

_IpeImage :: forall r. Prism' (IpeObject r) (IpeObject' Image r)

_IpeTextLabel :: forall r. Prism' (IpeObject r) (IpeObject' TextLabel r)

_IpeMiniPage :: forall r. Prism' (IpeObject r) (IpeObject' MiniPage r)

_IpeUse :: forall r. Prism' (IpeObject r) (IpeObject' IpeSymbol r)

_IpePath :: forall r. Prism' (IpeObject r) (IpeObject' Path r)

type IpeObject' g r = g r :+ IpeAttributes g r

An IpeObject' is essentially the oject ogether with its attributes

ipeObject' :: ToObject i => i r -> IpeAttributes i r -> IpeObject r

Shorthand for constructing ipeObjects

class ToObject i where

Methods

mkIpeObject :: IpeObject' i r -> IpeObject r

type IpeAttributes g r = Attributes' r (AttributesOf g)

type Attributes' r = Attributes (AttrMapSym1 r)

Attributes' :: * -> [AttributeUniverse] -> *

type family AttributesOf (t :: * -> *) :: [AttributeUniverse] where ...

Equations

AttributesOf Group = GroupAttributes

AttributesOf Image = CommonAttributes

AttributesOf TextLabel = CommonAttributes

AttributesOf MiniPage = CommonAttributes

AttributesOf IpeSymbol = SymbolAttributes

AttributesOf Path = PathAttributes

type family AttrMap (r :: *) (l :: AttributeUniverse) :: * where ...

The mapping between the labels of the the attributes and the types of the attributes with these labels. For example, the Matrix label/attribute should have a value of type 'Matrix 3 3 r'.

Equations

AttrMap r 'Layer = LayerName

AttrMap r Matrix = Matrix 3 3 r

AttrMap r Pin = PinType

AttrMap r Transformations = TransformationTypes

AttrMap r Stroke = IpeColor r

AttrMap r Pen = IpePen r

AttrMap r Fill = IpeColor r

AttrMap r Size = IpeSize r

AttrMap r Dash = IpeDash r

AttrMap r LineCap = Int

AttrMap r LineJoin = Int

AttrMap r FillRule = FillType

AttrMap r Arrow = IpeArrow r

AttrMap r RArrow = IpeArrow r

AttrMap r Opacity = IpeOpacity

AttrMap r Tiling = IpeTiling

AttrMap r Gradient = IpeGradient

AttrMap r Clip = Path r

data AttrMapSym1 (r6989586621679329293 :: Type) :: (~>) AttributeUniverse Type

attributes :: Lens' (IpeObject' g r) (IpeAttributes g r)

traverseIpeAttrs :: (Applicative f, AllConstrained TraverseIpeAttr (AttributesOf g)) => proxy g -> (r -> f s) -> IpeAttributes g r -> f (IpeAttributes g s)

traverse for ipe attributes

commonAttributes :: Lens' (IpeObject r) (Attributes (AttrMapSym1 r) CommonAttributes)

flattenGroups :: [IpeObject r] -> [IpeObject r]

collect all non-group objects

data View

The definition of a view make active layer into an index ?

Constructors

View [LayerName] LayerName

layerNames :: Lens' View [LayerName]

activeLayer :: Lens' View LayerName

data IpeStyle

for now we pretty much ignore these

Constructors

IpeStyle (Maybe Text) (Node Text Text)

styleName :: Lens' IpeStyle (Maybe Text)

styleData :: Lens' IpeStyle (Node Text Text)

basicIpeStyle :: IpeStyle

data IpePreamble

The maybe string is the encoding

Constructors

IpePreamble (Maybe Text) Text

encoding :: Lens' IpePreamble (Maybe Text)

preambleData :: Lens' IpePreamble Text

type IpeBitmap = Text

data IpePage r

An IpePage is essentially a Group, together with a list of layers and a list of views.

Constructors

IpePage [LayerName] [View] [IpeObject r]

layers :: forall r. Lens' (IpePage r) [LayerName]

views :: forall r. Lens' (IpePage r) [View]

content :: forall r r. Lens (IpePage r) (IpePage r) [IpeObject r] [IpeObject r]

emptyPage :: IpePage r

Creates an empty page with one layer and view.

fromContent :: [IpeObject r] -> IpePage r

Creates a simple page with a single view.

onLayer :: LayerName -> Getting (Endo [IpeObject r]) [IpeObject r] (IpeObject r)

This allows you to filter the objects on some layer.

>>> let page = IpePage [] [] []
>>> page^..content.onLayer "myLayer"
[]

contentInView :: Word -> Getter (IpePage r) [IpeObject r]

Gets all objects that are visible in the given view.

Note that views are indexed starting from 0. If the page does not have any explicit view definitions, this function returns an empty list.

>>> let page = IpePage [] [] []
>>> page^.contentInView 0
[]

withDefaults :: IpePage r -> IpePage r

Makes sure that the page has at least one layer and at least one view, essentially matching the behaviour of ipe. In particular,

if the page does not have any layers, it creates a layer named "alpha", and

if the page does not have any views, it creates a view in which all layers are visible.

data IpeFile r

A complete ipe file

Constructors

IpeFile (Maybe IpePreamble) [IpeStyle] (NonEmpty (IpePage r))

preamble :: forall r. Lens' (IpeFile r) (Maybe IpePreamble)

styles :: forall r. Lens' (IpeFile r) [IpeStyle]

pages :: forall r r. Lens (IpeFile r) (IpeFile r) (NonEmpty (IpePage r)) (NonEmpty (IpePage r))

ipeFile :: NonEmpty (IpePage r) -> IpeFile r

Convenience constructor for creating an ipe file without preamble and with the default stylesheet.

singlePageFile :: IpePage r -> IpeFile r

Convenience function to construct an ipe file consisting of a single page.

singlePageFromContent :: [IpeObject r] -> IpeFile r

Create a single page ipe file from a list of IpeObjects