Data.Geometry.Ipe.Writer

Description

Documentation

writeIpeFile :: IpeWriteText r => FilePath -> IpeFile r -> IO ()

Given a prism to convert something of type g into an ipe file, a file path, and a g. Convert the geometry and write it to file.

Write an IpeFiele to file.

writeIpeFile' :: IpeWrite t => t -> FilePath -> IO ()

Convert to ipe XML and write the output to a file.

writeIpePage :: IpeWriteText r => FilePath -> IpePage r -> IO ()

Creates a single page ipe file with the given page

toIpeXML :: IpeWrite t => t -> Maybe ByteString

Convert to Ipe xml

printAsIpeSelection :: IpeWrite t => t -> IO ()

Convert the input to ipeXml, and prints it to standard out in such a way that the copied text can be pasted into ipe as a geometry object.

toIpeSelectionXML :: IpeWrite t => t -> Maybe ByteString

Convert input into an ipe selection.

class IpeWrite t where

Types that correspond to an XML Element. All instances should produce an Element. If the type should produce a Node with the Text constructor, use the IpeWriteText typeclass instead.

Methods

ipeWrite :: t -> Maybe (Node Text Text)

class IpeWriteText t where

For types that can produce a text value

Methods

ipeWriteText :: t -> Maybe Text

ipeWriteAttrs :: (RecordToList rs, RMap rs, ReifyConstraint IpeWriteText (Attr f) rs, AllConstrained IpeAttrName rs, RecAll (Attr f) rs IpeWriteText) => Attributes f rs -> [(Text, Text)]

Functon to write all attributes in a Rec

writeAttrValues :: (RMap rs, ReifyConstraint IpeWriteText f rs, RecAll f rs IpeWriteText) => Rec f rs -> Rec (Const (Maybe Text)) rs

Writing the attribute values