Data.PlaneGraph.Draw

Documentation

drawPlaneGraph :: forall s v e f r. IpeOut (PlaneGraph s v e f r) Group r

Draws a planegraph using Marks, LineSegments, and simple polygons for vertices, edges, and faces, respectively. Uses the default IpeOuts to draw these elements.

drawPlaneGraphWith :: (Point 2 r -> v -> Maybe (IpeObject r)) -> (LineSegment 2 v r -> e -> Maybe (IpeObject r)) -> (SimplePolygon v r -> f -> Maybe (IpeObject r)) -> IpeOut (PlaneGraph s v e f r) Group r

Draws a planegraph using Marks, LineSegments, and simple polygons for vertices, edges, and faces, respectively.

genericDrawPlaneGraphWith :: ((VertexId' s :+ v) -> IpeObject r) -> ((Dart s :+ e) -> IpeObject r) -> ((FaceId' s :+ f) -> IpeObject r) -> IpeOut (PlaneGraph s v e f r) Group r

Draw a planegraph using the given functions. Fully generic in how we draw the objects.