Ipe.Attributes

Description

Possible Attributes we can assign to items in an Ipe file

Documentation

data AttributeUniverse

The possible Attributes supported in Ipe. To use these attributes, you'll likely need their Singletons's version which is Prefixed by an S. E.g. the Fill attribute is represented by a singleton 'SFill :: Sing Fill'.

Constructors

Layer

Matrix

Pin

Transformations

Stroke

Fill

Pen

Size

Dash

LineCap

LineJoin

FillRule

Arrow

RArrow

StrokeOpacity

Opacity

Tiling

Gradient

Clip

type LayerSym0 = 'Layer :: AttributeUniverse

type MatrixSym0 = 'Matrix :: AttributeUniverse

type PinSym0 = 'Pin :: AttributeUniverse

type TransformationsSym0 = 'Transformations :: AttributeUniverse

type StrokeSym0 = 'Stroke :: AttributeUniverse

type FillSym0 = 'Fill :: AttributeUniverse

type PenSym0 = 'Pen :: AttributeUniverse

type SizeSym0 = 'Size :: AttributeUniverse

type DashSym0 = 'Dash :: AttributeUniverse

type LineCapSym0 = 'LineCap :: AttributeUniverse

type LineJoinSym0 = 'LineJoin :: AttributeUniverse

type FillRuleSym0 = 'FillRule :: AttributeUniverse

type ArrowSym0 = 'Arrow :: AttributeUniverse

type RArrowSym0 = 'RArrow :: AttributeUniverse

type StrokeOpacitySym0 = 'StrokeOpacity :: AttributeUniverse

type OpacitySym0 = 'Opacity :: AttributeUniverse

type TilingSym0 = 'Tiling :: AttributeUniverse

type GradientSym0 = 'Gradient :: AttributeUniverse

type ClipSym0 = 'Clip :: AttributeUniverse

data SAttributeUniverse z where

Constructors

SLayer :: SAttributeUniverse ('Layer :: AttributeUniverse)

SMatrix :: SAttributeUniverse ('Matrix :: AttributeUniverse)

SPin :: SAttributeUniverse ('Pin :: AttributeUniverse)

STransformations :: SAttributeUniverse ('Transformations :: AttributeUniverse)

SStroke :: SAttributeUniverse ('Stroke :: AttributeUniverse)

SFill :: SAttributeUniverse ('Fill :: AttributeUniverse)

SPen :: SAttributeUniverse ('Pen :: AttributeUniverse)

SSize :: SAttributeUniverse ('Size :: AttributeUniverse)

SDash :: SAttributeUniverse ('Dash :: AttributeUniverse)

SLineCap :: SAttributeUniverse ('LineCap :: AttributeUniverse)

SLineJoin :: SAttributeUniverse ('LineJoin :: AttributeUniverse)

SFillRule :: SAttributeUniverse ('FillRule :: AttributeUniverse)

SArrow :: SAttributeUniverse ('Arrow :: AttributeUniverse)

SRArrow :: SAttributeUniverse ('RArrow :: AttributeUniverse)

SStrokeOpacity :: SAttributeUniverse ('StrokeOpacity :: AttributeUniverse)

SOpacity :: SAttributeUniverse ('Opacity :: AttributeUniverse)

STiling :: SAttributeUniverse ('Tiling :: AttributeUniverse)

SGradient :: SAttributeUniverse ('Gradient :: AttributeUniverse)

SClip :: SAttributeUniverse ('Clip :: AttributeUniverse)

type CommonAttributes = [Layer, Matrix, Pin, Transformations]

IpeObjects may have attributes. Essentially attributes are (key,value) pairs. The key is some name. Which attributes an object can have depends on the type of the object. However, all ipe objects support the Common Attributes

type TextLabelAttributes = CommonAttributes

All attributes applicable to TextLabels

type MiniPageAttributes = CommonAttributes

All attributes applicable to Minipages

type ImageAttributes = CommonAttributes

All attributes applicable to Images

type SymbolAttributes = CommonAttributes ++ [Stroke, Fill, Pen, Size]

All attributes applicable to Symbols/Marks

type PathAttributes = CommonAttributes ++ [Stroke, Fill, Dash, Pen, LineCap, LineJoin, FillRule, Arrow, RArrow, StrokeOpacity, Opacity, Tiling, Gradient]

All attributes applicable to Paths

type GroupAttributes = CommonAttributes ++ '['Clip]

All attributes applicable to Groups

A single attribute Attr

newtype Attr (f :: TyFun u * -> *) (label :: u)

Attr implements the mapping from labels to types as specified by the (symbol representing) the type family f

Constructors

GAttr

Fields

_getAttr :: Maybe (Apply f label)

getAttr :: forall f label f label. Iso (Attr f label) (Attr f label) (Maybe (Apply f label)) (Maybe (Apply f label))

pattern Attr :: Apply f label -> Attr f label

Constructor for constructing an Attr given an actual value.

pattern NoAttr :: Attr f label

An Attribute that is not set

traverseAttr :: Applicative h => (Apply f label -> h (Apply g label)) -> Attr f label -> h (Attr g label)

Traverse an attribute.

pureAttr :: (Applicative h, Apply f a ~ Apply g a) => Attr f a -> h (Attr g a)

Traverse for the situation where the type is not actually parameterized.

Attributes

newtype Attributes (f :: TyFun u * -> *) (ats :: [u])

A collection of Attributes.

Constructors

Attrs (Rec (Attr f) ats)

unAttrs :: Lens (Attributes f ats) (Attributes f' ats') (Rec (Attr f) ats) (Rec (Attr f') ats')

Get a vinyl Record with Attrs

traverseAttrs :: Applicative h => (forall label. Attr f label -> h (Attr g label)) -> Attributes f ats -> h (Attributes g ats)

Traverse implementation for Attrs

zipRecsWith :: (forall a. f a -> g a -> h a) -> Rec f as -> Rec g as -> Rec h as

Zip two Recs with the given function.

ixAttr :: forall at ats proxy f. at ∈ ats => proxy at -> Lens' (Attributes f ats) (Maybe (Apply f at))

Lens into a specific attribute, if it is set.

_Attr :: forall at ats proxy f. (at ∈ ats, RecApplicative ats) => proxy at -> Prism' (Attributes f ats) (Apply f at)

Prism into a particular attribute.

lookupAttr :: at ∈ ats => proxy at -> Attributes f ats -> Maybe (Apply f at)

Looks up a particular attribute.

setAttr :: forall proxy at ats f. at ∈ ats => proxy at -> Apply f at -> Attributes f ats -> Attributes f ats

Sets a particular attribute

takeAttr :: forall proxy at ats f. at ∈ ats => proxy at -> Attributes f ats -> (Maybe (Apply f at), Attributes f ats)

gets and removes the attribute from Attributes

unSetAttr :: forall proxy at ats f. at ∈ ats => proxy at -> Attributes f ats -> Attributes f ats

unsets/Removes an attribute

attr :: (at ∈ ats, RecApplicative ats) => proxy at -> Apply f at -> Attributes f ats

Creates a singleton attribute

Implementations for Common Attributes

data PinType

Possible values for Pin

Constructors

No

Yes

Horizontal

Vertical

data TransformationTypes

Possible values for Transformation

Constructors

Affine

Rigid

Translations

Text Attributes

Symbol Attributes

newtype IpeSize r

The optional Attributes for a symbol data SymbolAttributeUniverse = SymbolStroke | SymbolFill | SymbolPen | Size deriving (Show,Eq)

Size

Constructors

IpeSize (IpeValue r)

newtype IpePen r

Pen/Thickness

Constructors

IpePen (IpeValue r)

Path Attributes

data IpeDash r

Possible values for Dash

Constructors

DashNamed Text

DashPattern [r] r

data FillType

Allowed Fill types

Constructors

Wind

EOFill

type IpeOpacity = Text

IpeOpacity, IpeTyling, and IpeGradient are all symbolic values

type IpeTiling = Text

type IpeGradient = Text

data IpeArrow r

Possible values for an ipe arrow

Constructors

IpeArrow

Fields

_arrowName :: Text

_arrowSize :: IpeSize r

arrowSize :: forall r r. Lens (IpeArrow r) (IpeArrow r) (IpeSize r) (IpeSize r)

arrowName :: forall r. Lens' (IpeArrow r) Text

normalArrow :: IpeArrow r

A normal arrow

Group Attributes

Attribute names in Ipe

class IpeAttrName (a :: AttributeUniverse) where

For the types representing attribute values we can get the name/key to use when serializing to ipe.

Methods

attrName :: proxy a -> Text

writeAttrNames :: AllConstrained IpeAttrName rs => Rec f rs -> Rec (Const Text) rs

Writing Attribute names