Ipe.IpeRender

Description

Use iperender to generate png, pdf, or svg files.

Note that all functions in this module require that iperender is installed (it is bundled with ipe) and available on the path.

Documentation

ipeRenderWith

Arguments

:: Options
the options to use

-> FileType
output file type

-> FilePath
input file path

-> FilePath
output file path

-> IO ()

Call iperender to produce an image of the specified type.

note that pdf files produces with iperender cannot be opened with ipe.

ipeRender :: FileType -> FilePath -> FilePath -> IO ()

Call iperender with the default options.

note that pdf files produces with iperender cannot be opened with ipe.

data FileType

Output filetypes supported by iperender

Constructors

PNG

EPS

PDF

SVG

data Options

Options for iperender

Constructors

Options

Fields

pageNumber :: Int

viewNumber :: Int

resolution :: Int

transparent :: Background

crop :: Crop

defaultOptions :: Options

The default options in Ipe

data Background

Whether or not to render a transparent background in output png images.

Constructors

OpaqueBackground

TransparentBackground

data Crop

Whether or not to crop the output image.

Constructors

NoCrop

Crop